requirements.txt
astroid==2.3.3
autopep8==1.5.1
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.14.0
click==7.1.1
dnspython==1.16.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-WTF==0.14.3
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.19
packaging==20.3
Pillow==7.0.0
protobuf==3.6.1
pycodestyle==2.5.0
pycparser==2.20
pylint==2.4.4
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
six==1.14.0
tornado==6.0.4
typing-extensions==3.7.4.1
Werkzeug==1.0.0
wincertstore==0.2
wrapt==1.11.2
WTForms==2.2.1
